Generate initial board state on display. Add player.rb
Made changes so that the board is initialized with lots of pieces
objects representing a chess initial board state. These objects
each contain their symbol, being different depending on the color.

Each piece object also contains an array of move vectors that is
still pending of being developed. Method placed as a placeholder.