Add moves array and hashes. Add moves validator.
The pieces_gen now contain moves array or hashes depending on the
piece type. Pieces that can self-generate all their moves via recursion
have an array, wereas those that can't contain a hash.

The array contains all the possible paths wereas the hash contains moves
and rules.

valid_move? is a private method contained in player that checks wether
a move is valid or not depending on all the chess rules. Implementation
pending.