Fix copy misbehavior inside #enables_check?
The player's class method, #enables_check? had an error that
caused the whole application to misbehave. The board was modified
within this method and the whole #valid_move method was bypassed.

This change ensures the application behaves correctly and that
the changes being displayed were made within the main #move_piece
method.